requirements.txt
boto3==1.34.30
psycopg2-binary==2.9.9
pydantic==2.10.6
python-dotenv==1.0.1
PyYAML==6.0.1